+++ sqlmesh_dag_generator/__init__.py
@@ -2,11 +2,16 @@
 SQLMesh DAG Generator - Open Source Airflow Integration for SQLMesh
 """
 
-__version__ = "0.9.13"
+__version__ = "0.9.14"
 
 
 from sqlmesh_dag_generator.generator import SQLMeshDAGGenerator
 from sqlmesh_dag_generator.config import DAGGeneratorConfig, RecoveryConfig
+from sqlmesh_dag_generator.triggers import (
+    ModelTriggerConfig,
+    parse_trigger_from_tags,
+    resolve_model_trigger,
+)
 from sqlmesh_dag_generator.airflow_utils import (
     resolve_credentials,
     register_credential_resolver,
@@ -27,6 +32,9 @@ __all__ = [
     "SQLMeshDAGGenerator",
     "DAGGeneratorConfig",
     "RecoveryConfig",
+    "ModelTriggerConfig",
+    "parse_trigger_from_tags",
+    "resolve_model_trigger",
     "resolve_credentials",
     "register_credential_resolver",
     "CredentialResolver",